requirements.txt
about-time==4.2.1
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.15.1
alive-progress==3.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
astroid==3.3.8
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==24.3.0
autopep8==2.3.1
babel==2.17.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.3.0
blinker==1.9.0
blis==1.2.0
bokeh==3.7.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chamfer_3D==0.0.0
chardet==5.2.0
charset-normalizer==3.4.1
chumpy @ git+https://github.com/mattloper/chumpy@51d5afd92a8ded3637553be8cef41f328a1c863a
clarabel==0.10.0
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
cmeel==0.57.1
coacd==1.0.5
colorama==0.4.6
comet-ml==3.32.8
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7
configobj==5.0.9
contourpy==1.2.1
cvxopt==1.3.2
cycler==0.12.1
cymem==2.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-sdk==0.47.0
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.32.1
dill==0.3.9
docker==7.1.0
docker-pycreds==0.4.0
docopt==0.6.2
dulwich==0.22.7
easydict==1.13
einops==0.8.0
etils==1.12.0
everett==3.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.12
fasteners==0.19
fastjsonschema==2.21.1
ffmpeg==1.4
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
fonttools==4.55.3
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
gdown==5.2.0
genesis-world==0.2.1
gitdb==4.0.12
GitPython==3.1.44
glcontext==3.0.0
glfw==2.8.0
google-auth==2.38.0
grapheme==0.6.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.69.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.5
higher==0.2.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
imgui==2.0.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
inform==1.33
iopath==0.1.10
ipdb==0.13.13
ipykernel==7.1.0
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jpeg4py==0.1.4
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.8
langcodes==3.5.0
language_data==1.3.0
libigl==2.5.1
lightning==2.5.0.post0
lightning-utilities==0.11.9
line_profiler==4.2.0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.1
Mako==1.3.9
manopth @ file:///home/zxc417/Projects/reproductions/manopth
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mesh-to-sdf==0.0.15
mistune==3.1.4
ml_dtypes==0.5.0
mlflow==2.21.0
mlflow-skinny==2.21.0
moderngl==5.12.0
moderngl-window==3.0.3
moviepy==2.1.2
mpmath==1.3.0
mujoco==3.2.5
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netron==8.2.5
networkx==3.4.2
notebook==7.4.7
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
numpy-quaternion==2024.0.7
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
onnxscript==0.1.0.dev20241219
opacus==1.5.2
open3d==0.19.0
open3d-cpu==0.19.0
opencv-contrib-python==4.10.0.84
opencv-contrib-python-headless==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
OpenEXR==3.3.2
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opt_einsum==3.4.0
overrides==7.7.0
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pi==0.1.2
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plyfile==1.1
pointnet2==0.0.0
pooch==1.8.2
portalocker==3.0.0
preshed==3.0.9
proglog==0.1.10
progress==1.6
prometheus_client==0.23.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
proxsuite==0.7.1
psutil==6.1.1
ptyprocess==0.7.0
PuLP==2.9.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybullet==3.2.7
pycodestyle==2.12.1
pycollada==0.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
pyg-lib @ file:///home/zxc417/Downloads/Lib/pyg_lib-0.4.0%2Bpt25cu124-cp310-cp310-linux_x86_64.whl#sha256=b620ac6ae03b9933f6715f5fb8228d062da42c5b511f8512afe2d79029bbc43d
PyGEL3D==0.6.1
pygit2==1.7.0
pyglet==2.0.20
PyGLM==2.7.3
pygltflib==1.16.0
Pygments==2.18.0
pylint==3.3.3
pymeshlab==2023.12.post2
PyOpenGL==3.1.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PyQt6==6.7.1
PyQt6_sip @ file:///croot/pyqt-split_1740498191142/work/pyqt_sip
pyquaternion==0.9.9
pyrender==0.1.45
pysdf==0.1.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==4.0.0
pytorch-lightning==2.0.0
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@055ab3a2e3e611dff66fa82f632e62a315f3b5e7
pytz==2024.2
pyvista==0.44.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.3.0
qpsolvers==4.4.0
quantiphy==2.20
rdkit==2024.9.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roma==1.5.1
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
safetensors==0.5.0
scalene==1.4.1
scikit-image==0.19.3
scikit-learn==1.6.0
scikit-video==1.1.11
scipy==1.15.2
scooby==0.10.0
screeninfo==0.8.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==5.0.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
shortuuid==1.0.13
simplejson==3.19.3
sip @ file:///croot/sip_1738856193618/work
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
smplx==0.1.28
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
taichi==1.7.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
tetgen==0.6.4
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2024.12.12
tinycss2==1.4.0
tokenizers==0.21.2
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
tomlkit==0.13.2
torch==2.5.1
torch-geometric==2.6.1
torch_cluster @ file:///home/zxc417/Downloads/Lib/torch_cluster-1.6.3%2Bpt25cu124-cp310-cp310-linux_x86_64.whl#sha256=12c2b7f800ea84545dfbb42aa5a1acefbb274b31aecb20d0cb18431a16270463
torchaudio==2.5.1
torchgeometry==0.1.2
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
transforms3d==0.4.2
trimesh==4.6.1
triton==3.1.0
tvm==1.0.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
usd-core==24.11
uvicorn==0.34.0
vtk==9.3.1
wandb==0.19.4
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==14.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.0
wurlitzer==3.1.1
xyzservices==2025.1.0
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
